+++ application/controllers/Ajax/GetKey.php
@@ -32,4 +32,20 @@ class GetKey extends AJAX_Controller {
 		}
 	}
 
+	public function restore() : void {
+		if($this->ion_auth->logged_in()) {
+			if(!$this->limiter->limit('new_api_key', 10)) {
+				$api_key = $this->User->restore_api_key();
+				$json    = ['api-key' => $api_key];
+
+				$this->output
+					->set_content_type('application/json')
+					->set_output(json_encode($json));
+			} else {
+				$this->output->set_status_header('429', 'Rate limit reached.');
+			}
+		} else {
+			$this->output->set_status_header('400', 'Not logged in.');
+		}
+	}
 }